+++ test/e2e010/test_upgrade_0.11.go
@@ -31,7 +31,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/wait"
 )
 
-var target011Version string = "0.11.0"
+var target011Version string = "0.11.1"
 
 func astarteUpgradeTo011Test(f *framework.Framework, ctx *framework.Context) error {
 	namespace, err := ctx.GetWatchNamespace()